added Timer GUI (list, configuration dialog) to select and show randomizer value updated database methods for randomizer value updated next alarm time generation methods to make use of randomizer value

+++ Smartphone/src/main/java/eu/power_switch/database/table/phone/call/CallEventTable.java
@@ -47,29 +47,9 @@ public class CallEventTable {
     }
 
     public static void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        switch (oldVersion) {
-            case 1:
-            case 2:
-            case 3:
-            case 4:
-            case 5:
-            case 6:
-            case 7:
-            case 8:
-            case 9:
-            case 10:
-            case 11:
-            case 12:
-            case 13:
-            case 14:
-            case 15:
-            case 16:
-            case 17:
-            case 18:
-                db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-                onCreate(db);
-                break;
-
+        if (oldVersion <= 19) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(db);
         }
     }
 }
